Got Solenoid objects instanciated and pouring. Also have Header.h file to kinda organize stuff

+++ src/Solenoid.h
@@ -1,10 +1,11 @@
 #ifndef SOLENOID_H
 #define SOLENOID_H
+#include "Header.h"
 
 class Solenoid
 {
 private:
-    short int activationNum;  //The number that gets fed to the shift registers to activate this specific solenoid's relay (to pour)
+    int activationNum;  //The number that gets fed to the shift registers to activate this specific solenoid's relay (to pour)
     short int xPos;           //The number representing the fixed location of the solenoid. "How many revolutions away from the stepper motor am I"
     
 
@@ -13,17 +14,17 @@ public:
     Solenoid();
 
     //Parameterized constructor
-    Solenoid(short int arg_activationNum, short int arg_xPos);
+    Solenoid(int arg_activationNum, short int arg_xPos);
 
     // Opens solenoid (pours) for X milliseconds.
     void Pour(int milliseconds);
 
     // Getters
     short int getxPos();
-    short int getactivationNum();
+    int getactivationNum();
 
     // Setters
     void setxPos(short int arg_xPos);
-    void setactivationNum(short int arg_activationNum);
+    void setactivationNum(int arg_activationNum);
 };
 #endif
